feat: esc10b composition
