+++ cmd/api/src/analysis/ad/adcs_integration_test.go
@@ -1279,8 +1279,6 @@ func TestADCSESC6a(t *testing.T) {
 					operation.Operation.SubmitReader(func(ctx context.Context, tx graph.Transaction, outC chan<- analysis.CreatePostRelationshipJob) error {
 						if err := ad2.PostADCSESC6a(ctx, tx, outC, groupExpansions, innerEnterpriseCA, innerDomain, cache); err != nil {
 							t.Logf("failed post processing for %s: %v", ad.ADCSESC6a.String(), err)
-						} else {
-							return nil
 						}
 
 						return nil
