+++ packages/go/analysis/ad/adcs.go
@@ -145,7 +145,7 @@ func processEnterpriseCAWithValidCertChainToDomain(enterpriseCA, domain *graph.N
 
 	operation.Operation.SubmitReader(func(ctx context.Context, tx graph.Transaction, outC chan<- analysis.CreatePostRelationshipJob) error {
 		if err := PostADCSESC6b(ctx, tx, outC, groupExpansions, enterpriseCA, domain, cache); err != nil {
-			log.Errorf("failed post processing for %s: %v", ad.ADCSESC6a.String(), err)
+			log.Errorf("failed post processing for %s: %v", ad.ADCSESC6b.String(), err)
 		}
 		return nil
 	})
