+++ packages/javascript/bh-shared-ui/src/components/HelpTexts/ADCSESC10a/ADCSESC10a.tsx
@@ -19,7 +19,7 @@ import WindowsAbuse from './WindowsAbuse';
 import LinuxAbuse from './LinuxAbuse';
 import Opsec from './Opsec';
 import References from './References';
-import Composition from "./Composition";
+import Composition from './Composition';
 
 const ADCSESC10a = {
     general: General,
@@ -27,7 +27,7 @@ const ADCSESC10a = {
     linuxAbuse: LinuxAbuse,
     opsec: Opsec,
     references: References,
-    composition: Composition
+    composition: Composition,
 };
 
 export default ADCSESC10a;
